// packages/shared/generated/api/hire_me/Salary.ts
// @generated
// This file is automatically generated by Kanel. Do not modify manually.

import { NumRange } from "../../../types/api/Ranges.js";
import type { RoleId } from "./Role.js";
import type { default as SalaryPeriod } from "./SalaryPeriod.js";

/** Identifier type for hire_me.salary */
export type SalaryId = number & { __brand: "SalaryId" };

/** Represents the table hire_me.salary */
export default interface Salary {
	id: SalaryId;

	role_id: RoleId;

	salary_range: NumRange;

	salary_includes_super: boolean;

	salary_period: SalaryPeriod;

	salary_currency: string;
}

/** Represents the initializer for the table hire_me.salary */
export interface SalaryInitializer {
	role_id: RoleId;

	salary_range: NumRange;

	salary_includes_super: boolean;

	salary_period: SalaryPeriod;

	salary_currency: string;
}

/** Represents the mutator for the table hire_me.salary */
export interface SalaryMutator {
	role_id?: RoleId;

	salary_range?: NumRange;

	salary_includes_super?: boolean;

	salary_period?: SalaryPeriod;

	salary_currency?: string;
}